Seperated out model/interface components for CacheService
