+++ SimcProfileParser/Model/DataSync/SimcFileTypeEnum.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SimcProfileParser.Model.DataSync
+{
+    public enum SimcFileType
+    {
+        ScSpellDataInc,
+        ItemDataInc,
+        RandomPropPointsInc,
+        ScaleDataInc,
+        ScEnumsHpp,
+        DataEnumsHh
+    }
+}
